Merge pull request #480 from RosettaCommons/vmullig/demo_fix_1
Fixing a "bug" for the Rosetta 201 training session.



This is intended to demo the GitHub pull request workflow for Bakerites during the Rosetta 201 lecture on Thursday. The "bug" is a typo in an output message in GeneralizedKIC that I introduced deliberately a few days ago. Fixing this should result in cosmetic integration test changes in the GeneralizedKIC tests.

(This is the "Martha Stewart" pull request that I'm setting up the day before so that it will be all tested and ready to go during the actual session, so that I can push the big green button there. "So put that in the oven and let it bake. Now here's one we made earlier.")

Old repository SHA1: 2b13761214ace18a200138d589dc3f51f2f6b897