Update default-torsions.txt
The C-N-CA-CB and N-C-CA-CB improper torsions for CYS were being scored, but the corresponding torsions for CYD were not being scored. I confirmed with Frank that this is an error, one that I wished to fix before handling the whole disulfide patch merge--this is a simple enough bugfix that it doesn't require a pull request.

Old repository SHA1: e03980aa1f47dadfc8b0b761ab5f50c8476d4abb