Merge pull request #519 from RosettaCommons/aleaverfay/beauty
This is a pre-commit before beautification takes place. It adds semi-colons behind some macro invocations (there shouldn't even be macro invocations!) so that the beautifier can run smoothly.

It generates a number of cosmetic integration test changes as I've replaced a tab inside a string with some spaces in score.cc. If you look at the file, you see that every other output line in the "how to run this program" section is indented with spaces and so having a tab in there just distorts the output text alignment.

The commit also changes the DOS line-endings to unix format for a handful of files.

Old repository SHA1: 317d3adb1fdbf62830d21790f48ccfd9704be4d6