Merge pull request #558 from RosettaCommons/dougrenfrew/tcre_energy_methods_clean
Adding a two-component reference energy (TCRE, a.k.a. split unfolded) term. It is intended to serve as a replacement for the unfolded term and should be applicable for the design of alpha/beta/gamma-peptides, peptoids, OOPs, HBSs, and other NCBBs as it is evaluated based on atom type composition.

Riley and Andy have contributed this greatly.

Old repository SHA1: c62827e122f24ef1ffe5f4655b26ea7aea78d540

Co-authored-by: P. Douglas Renfrew <renfrew@nyu.edu>
Co-authored-by: Andy Watkins <amw579@nyu.edu>
Co-authored-by: Andrew Watkins <amw579@nyu.edu>