Merge pull request #633 from RosettaCommons/rfalford12/add_mprelax_sfxn_constructor
Add constructor to specify scoring function in MPRelax
-----------------------------------------------------------------------------
Added new constructor to MPRelax that takes a ScoreFunction. The ScoreFunction is also now a private member of the class to accommodate. Currently intended for use in PyRosetta so not hooked in anywhere

Old repository SHA1: b3b878070d67533736bd055bff5126e80877260b