Merge pull request #628 from RosettaCommons/kylebarlow/general_dev
Fixing -Wpotentially-evaluated-expression warnings-as-errors added in Clang-3.6

Some of these occurred in the cppdb external code. I first tried updating this code to the newest released version, but this didn't fix the warnings, and also broke a pilot app, so I reverted that change.

I suppressed the warnings in the external code, as we do not maintain it, and it doesn't look like a warning that really matters anyways.

In the Rosetta code, I think the warning was being caused by a deference of an owning pointer not playing well with typeid(), so I fixed the warning by first creating a temporary reference to typeid(). We seem to be good OOPers for the most part, so there aren't many uses of typeid() in the codebase to bring this up again.

Old repository SHA1: 12782030840e8802817069872c99c18723aa52dd