Merge pull request #684 from RosettaCommons/jadolfbr/proto_cdr4
DE Loop/Proto CDR4 analysis for the Antibody Features Reporters
-----------------------------------------------------------------------------------------

Much of this work, especially in terms of the Features R framework was done by Simon Kelow (@skelow).

This merge enables the Antibody Features Reporter and associated R scripts to analyze the DE loop of antibodies, what we are also calling proto_cdr4. Both names are not ideal, but for now, this is the way they are.

A few functions in AntibodyInfo were refactored in order to return information on this loop as if it was a CDR. This will enable the full antibody framework to work with this loop, and has allowed us to use the implemented feature reporters without too many changes.

A set of antibody-specific options have been added to the features R framework to enable global use of these loops during antibody analysis. Documentation on the wiki will be updated as this merge happens.

No integration or unit test changes seen except the currently failing ones.

Old repository SHA1: 010540855af75accca97020229704a1e048fbcc1

Co-authored-by: Simon Kelow <simon.kelow@gmail.com>
Co-authored-by: jadolfbr <jadolfbr@gmail.com>
Co-authored-by: Simon <simon.kelow@gmail.com>