Merge pull request #695 from RosettaCommons/rfalford12/remove_old_rm_code
Removing old resource manager hooks in the membrane framework
----------------------------------------------------------------------------------------------
When the membrane framework was in development, we added hooks to the resource manager. However, development stopped on these. Currently, the code does not compile, and if it did, it would not work with the current version of the RosettaMP framework. Once we update to JD3/resource manager, I'll try this again but probably from scratch. No reason for non-compiling, and non-functional files to complicate master.

Old repository SHA1: ce2f00cea247c0e3d6fab9f1f27d9b93b13f6753