Merge pull request #757 from RosettaCommons/vmullig/fix_invrotremodel
Fix the inverse_rotamer_remodel integration test

This seems to be failing due to a bunch of missing mm_std bond angle parameters. It bothers me, though, that this was NOT failing on the Linux builds. There was an option for ignoring missing bond angle parameters, but (a) this was set to false, and (b) even if this were set to true, you should end up with a segfault on any system if you ignore the missing parameters. Instead, the segfault was only happening on Mac. I don't know why missing bond angle parameters were being ignored, or why the segfault was only on Mac.

I've removed the option (so there should ALWAYS be an error thrown and an abort if there are missing bond angle parameters), and have added the missing bond angle parameters to the database (copying similar bond angle parameters and adding comments indicating that these are copies).

Old repository SHA1: 91491aae1e8c480edd78719564fde8a2f9ae8c8b