Merge pull request #764 from RosettaCommons/vmullig/fix_compile
Fixing the build.

Sergey Ovchinnikov just discovered that the existing build is actually a broken build, despite passing all build tests on the test server and on my machine. The problem is that I removed the sole member of an option group, so that that option file is no longer generated by the options.py script. There are a few files that include it, though. The file is there if it has ever been generated (e.g. in all test server builds, and on any existing checkout of Rosetta), so Rosetta compiles happily, but it is missing on new checkouts, so compilation fails there.

I test-built this after deleting all of the auto-generated options files and regenerating them. Will merge without testing, to get master building properly as quickly as possible.

Old repository SHA1: f32089824395b549e2284fc5b7123c02d2e45371