Merge pull request #773 from RosettaCommons/vmullig/muteall_issue
Fixing crash bug when "-mute all" option is used.

This seems to be an issue with using boost::format with the new tracer << macros when tracers are muted. The solution seems to be to enclose the tracer output in an "if(TR.visible()) {}" check. This problem was noticed in Remodel, though it might recur elsewhere.  To err to the side of caution, I've also bracketed a few other cases where boost::format is used in "if(TR.visible()) {}" checks.

All unit and integration tests pass  on my machine.  Merging without branch testing due to test server backlog.  Hopefully this is the last of the bugs currently plaguing master.

Old repository SHA1: 548b2739fd00841b22b05346b48330ae1b09c93e