Merge pull request #854 from RosettaCommons/everyday847/include-what-you-use
Now that we build under Windows/PyRosetta, we are ready to merge. There is a lot of work left to be done in this domain, but it will be done under separate pull requests--this one is already substantial in size. Basic/numeric/utility/core1/core2 are done (but could always use separate passes), as well as a couple public apps.

It is not fun to attempt this from a Mac, as the Linux builds are more persnickety. And I really wish I had the opportunity to locally test the Windows build--maybe, just maybe, I'll set that up before I try again.

Old repository SHA1: 034260ee58cd64cfa5de22cfb77526af259859ad