Merge pull request #880 from RosettaCommons/vmullig/reset_issue
Fix the functionality of RESET

The commutativity-breaking resfile command RESET, which is necessary for noncanonical design (since noncanonicals can only be turned ON while canonicals can only be turned off) wasn't resetting canonicals properly. This command now:

-    Disables all noncanonicals (as it did before).
-    Enables all canonicals (which it was not doing properly before).
-    Enables design and repacking.

Which is what the documentation claimed that it was doing.

I've also moved the RESET functionality to a single function in the ResidueLevelTask class, so that we don't have a lot of ways for external things to break commutativity. (Anything can call the reset() function, but it will always break commutativity in exactly this way.)

Old repository SHA1: 75b7c9466e48a2b13516989b00f17bc8a5191e45