Merge pull request #928 from RosettaCommons/vmullig/shuffle_libraries
Move ResidueSelectors to core.3 from core.4, and to a new directory

This is in anticipation of development work that will allow ResidueSelectors to be used as input to Constraints (and other things). I think it makes sense for ResidueSelectors to be moved away from the PackerTask machinery: though they can be used as input into a TaskOperation, they really should become the standard way of selecting residues for anything in Rosetta that takes a residue selection.

Tasks:

--    Move ResidueSelectors to core/select/residue_selectors. Everything associated with just selecting residues (or atoms, or anything else that we might one day be able to select), independent of the packer, will now live in core/select (core.3); packer-associated stuff will remain in core/pack (core.4).
--    Keep the ClashBasedRepackShellSelector in core/pack/task/residue_selector, since it has core.4 dependencies. Create a Creator for this, since its base class (and the Creators list associated with the base class) is now in a different namespace.
--    Move interface_vector_calculate from core/pack/task/operation/util to core/select/util, and from core.4 to core.3. This resolves library dependency issues and keeps selection logic separate from packer-specific logic.
--    Move SelectResiduesByLayer from core/util to core/select/util, and from core.4 to core.3. This also resolves library dependency issues and helps to group selection logic in core/select.
--    Ensure that all namespaces, all tracer names, and all unit tests are updated to reflect the new directory and namespace organization.

All of this should result in cosmetic integration test changes resulting from changed namespaces (which is why I want it to be a separate pull request from the other stuff that I'm doing in pull request #926), but no actual functionality changes.

Old repository SHA1: 153a0055353ac28b0e704dd3ffe835f0e898d5b0

Co-authored-by: Sergey Lyskov <sergey.lyskov@jhu.edu>