Merge pull request #985 from RosettaCommons/roccomoretti/user_database
Implement Jonathan Sheehan's suggestion of caching the Dunbrack library in the user's home directory.

Behavior in the standard case should be the same - the Dunbrack binary is read/written to the database directory. The big change is that when that directory is not writable, then the binary will be cached under the user's home directory. (This should be transparent for PyRosetta users.)

Priority of locations for binary search (if any of these locations is not set or cannot be written to, it will move on to the next one.):

*    Directory specified with -in::path::database_cache_dir option
*    Directory specified with ROSETTA3_DBCACHE environment variable
*    The same database directory from which the ASCII files were/would be read
*    A .rosetta/database/ directory in the user's home directory, determined from the first non-empty setting of
    *    The XDG_CONFIG_HOME environment variable
    *    The HOME environment variable
    *    The home directory specified in the POSIX password structure information.

If none of these are writable, then it will skip database file writing.

People who are working with multiple versions of the Dunbrack library may need to be slightly careful, as the location where the binary is read/written to may be different from where you expect. Use the validate_database application to check the cached binary, if you're seeing funky results.

Old repository SHA1: 29d47c943689abb7017c87b206a2dfc00e55c13f