Merge pull request #1088 from RosettaCommons/JWLabonte/sugars/rotamers
Post-XRW/Carbohydrates: Reducing number of rotamers for carbohydrate ResidueTypes

This merge will greatly reduce the number of rotamers generated for carbohydrate residues. Instead of sampling every 60 degrees, we will only sample every 120 degrees in the staggered orientations. Hopefully, this will prevent packing from running out of memory when dealing with poses with a large number of sugar residues. A better method of packing for sugars still needs to be developed, but this may help in the interim.

These changes will reduce the number of rotamers for a typical, unmodified aldohexopyranose, such as glucopyranose, from 1,296 to 81 rotamers. It will reduce the number of rotamers for neuraminic acid, a 9-carbon sugar, from 7,776 to 243 rotamers. These numbers will still be higher for modified sugars. For example, the common N-acetyl varieties will jump to 162 and 486 rotamers, respectively.

carbohydrate and vancomycin integration test changes expected, the latter showing a change from 1382 to 119 rotamers and 412 kb to 22 kb with only a 0.1 change in scoring.
All other tests pass.

Old repository SHA1: d2fca316d81abc46bd366c11118602e4c9fe2993