Again increasing the RMS cutoff for a splice test following PR #1079...
this time, because the (non-debug) integration tests are consistently
failing with an RMS of the spliced in segment of 0.54A.  It is surprising
that the splice tests are so sensitive, as PR #1079 only applies the slightest
bit more repulsion for hydrogens right as their interaction energy is about
to go to zero.

I have emailed with Gideon, and he is looking into whether something more
worrisome is going on.

Old repository SHA1: c4b2e27982b3cd850e7777c23e1d370f82919bd0