Merge pull request #673 from RosettaCommons/smlewis/pose_mover_file_itest
In the spirit of merging old PRs, here's an integration test I wrote on the van back from RosettaCON last year.  The comment from its command file explains its purpose:

This test is not an integration test in the traditional sense: it is not testing a run of the code.  Instead, this test just examines the code itself for a few important central classes.  These are important base classes used by everyone, and modifications to those classes should be approved by the community (or, let's be serious, Andrew).  Failing this integration test does NOT mean something is wrong - just that you've changed a file it watches and should be prepared to justify your change.

itests will break because a new itest is added.  No code changes.

Old repository SHA1: 68379846a2c029faa76690f8afd86e0a2ea1c73d