Merge pull request #1102 from RosettaCommons/vmullig/fix_helix_fit
Fixing issues with fit_helixparams app.

Some code in the numeric libraries for superimposing structures and calculating RMSDs is throwing off my fit_helixparams app. A small value (1.0e-7) is currently added to certain numbers to avoid a zero determinant in edge cases. This small value is resulting in a skew to the transformation matrices in some cases, and it turns out that it's because it's TOO small. I won't take it out, but I'll add a pass-through to let a specific protocol override it with a larger value. Initial tests indicate that this greatly improves the behaviour of the fit_helixparams app.

@hssnzdh2 @everyday847

Tasks:

-    Modify the rmsfitca2() function in numeric/model_quality/rms.cc/hh to take a double-precision float for the offset value. (Defaults to 1.0e-7 -- preserve the old behaviour).
-    Modify calling functions similarly.
-    Modify the fit_helixparams app to pass "1.0e-5" to rmsfitca2().
-    Add option for user to set the offset value to pass to rmsfitca2().
-    Beauty.

Old repository SHA1: 0f337566b202da1cb87e94bd01245d15d70e3cc7