Merge pull request #1107 from RosettaCommons/vmullig/rmstest
Adding integration test for the fit_helixparams app in a case that was producing bad geometry and bad fits prior to my corrections to the RMS machinery.

I want the problem to STAY fixed. The fit_helixparms_rms integration test should produce a PDB file in which every mainchain atom (except the carbonyl oxygen) has a copper atom that overlays perfectly on it. Distorted geometry or imperfect copper atom placement is a sign that something has gone wrong. Small changes to the coordinates that do NOT distort the geometry or affect the placement of copper atoms relative to backbone atoms are acceptable.

There is a README.txt file that explains this in the integration test directory. I'll also add the expected output to the expected_results directory.

Old repository SHA1: 2e381ac90bb3c8a1a0215ae230111a7e84ef72f2