Merge pull request #1140 from RosettaCommons/vmullig/fix_macrocycles
Trying to fix the macrocycles integration test

The app has a while( !closed ) loop that can potentially loop forever. I think that my changes didn't actually break the app -- they just altered the trajectory so that it no longer converges in the time limit for the integration test. This is a temporary workaround that causes the app to break out from that loop if a maximum number of iterations for loop closure is passed.

(Incidentally, please don't ever close a loop like this in a real instance! Use GenKIC, CCD, or anything else -- it'll definitely be faster!)

@everyday847

Old repository SHA1: c7374cd9a951e1311288596d6c5de326a5593d1f